now you can post some data via http when requesting a widget (#205)
## Problem

There is a use case in which we want to start a Beagle remote flow but the service needs some parameters. So we needed a way to instruct Beagle how to pass those parameters when making a request.

## Solution

Added a property on `ScreenType.Remote` called `additionalData: RemoteAdditionalData?` that can be used to pass anything. And if users pass an `HttpAdditionalData`, the default `NetworkClient` implementation will put those properties in the http request.

## Commits

* now you can post some data via http when requesting a widget

* use typealias to suggest proper usage

* abstracting logic of build an absolute url into a dependency

* fix BeagleDemo to work with new UrlBuilder